needed int instead of long

+++ src/main/java/com/is/buyabike/service/impl/ProductServiceImpl.java
@@ -42,6 +42,6 @@ public class ProductServiceImpl implements ProductService {
     @Transactional
     @Override
     public void deleteProduct(Product product) {
-        repo.delete(product.getId());
+        repo.delete((int) product.getId());
     }
 }
